refactor(dto): AttendanceHistoryByDate가 model에 더 적합하여 패키지 이동
