+++ src/main/java/model/AttendanceHistory.java
@@ -1,6 +1,5 @@
 package model;
 
-import dto.AttendanceHistoryByDate;
 import java.time.LocalDate;
 import java.util.Collections;
 import java.util.HashMap;
