+++ src/test/java/model/AttendanceHistoryTest.java
@@ -3,7 +3,6 @@ package model;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import dto.AttendanceHistoryByDate;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.List;
